pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brandkit"
version = "0.1.0"
description = "Brand name generator with phonaesthetic scoring and trademark checking"
readme = "README.md"
license = {text = "GPL-3.0"}
authors = [
    {name = "Joern Ihlenburg", email = "joern@ihlenburg.de"}
]
keywords = [
    "brand",
    "naming",
    "generator",
    "phonaesthetics",
    "trademark",
    "linguistics"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Linguistic",
]
requires-python = ">=3.10"
dependencies = [
    "PyYAML>=6.0",
    "requests>=2.28",
]

[project.optional-dependencies]
llm = [
    "anthropic>=0.18",
]
excel = [
    "openpyxl>=3.1",
]
ui = [
    "rich>=13.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
]
all = [
    "brandkit[llm,excel,ui,dev]",
]

[project.urls]
Homepage = "https://github.com/jihlenburg/brandkit"
Repository = "https://github.com/jihlenburg/brandkit"
Changelog = "https://github.com/jihlenburg/brandkit/blob/main/CHANGELOG.md"
Issues = "https://github.com/jihlenburg/brandkit/issues"

[project.scripts]
brandkit = "brandkit.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["brandkit*"]

[tool.setuptools.package-data]
"brandkit.generators.phonemes" = ["*.yaml"]
"brandkit.configs" = ["*.yaml"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"
